Warm the file cache on a background worker thread
The first request for an uncached file read and processed it
synchronously (Sass, Babel, Snockets, UglifyJS, gzip), blocking the
event loop and slowing the whole site. Extract the processing pipeline
into a shared processor module and run it in a background worker thread
that lazily reads every file in the directory, posting finished cache
entries back to the main thread. The synchronous read stays as the
per-request fallback, so behavior is unchanged for any file not yet
warmed. Warming is skipped while watching (it registers discovered
dependencies on the main thread) and when options carry functions
(which can't be cloned into the worker), to keep cached output identical
to the synchronous path.
