+++ lib/index.js
@@ -1,19 +1,98 @@
-const crypto = require('node:crypto');
-const fs = require('node:fs');
 const path = require('node:path');
-const zlib = require('node:zlib');
+const { Worker } = require('node:worker_threads');
 
-const babel = require('@babel/core');
 const chokidar = require('chokidar');
-const mime = require('mime');
 const Negotiator = require('negotiator');
-const sass = require('sass');
-const sassGraph = require('sass-graph');
-const Snockets = require('snockets');
-const UglifyJS = require('uglify-js');
-const UglifyCss = require('uglifycss');
 
-const gzipContentTypes = require('./gzipContentTypes.js');
+const { createProcessor, hashifyUrl, parseUrlPath } = require('./processor.js');
+
+/**
+ * Deeply checks whether a value contains any functions, which can't be
+ * structured-cloned to a worker thread. When options carry functions (e.g. a
+ * Babel plugin or Sass importer), the worker can't reproduce the main thread's
+ * output, so background warming is disabled to avoid caching divergent results.
+ * @param {*} value
+ * @param {Set} [seen]
+ */
+function hasFunction(value, seen = new Set()) {
+    if (typeof value === 'function') {
+        return true;
+    }
+
+    if (value && typeof value === 'object') {
+        if (seen.has(value)) {
+            return false;
+        }
+
+        seen.add(value);
+
+        for (const key of Object.keys(value)) {
+            if (hasFunction(value[key], seen)) {
+                return true;
+            }
+        }
+    }
+
+    return false;
+}
+
+/**
+ * Restores Buffers lost to structured cloning. Binary content posted from the
+ * worker arrives as a Uint8Array; wrap it back into a Buffer so it behaves like
+ * a synchronously-read file.
+ * @param {object} file
+ */
+function reviveFile(file) {
+    if (typeof file.content !== 'string') {
+        file.content = Buffer.from(file.content.buffer, file.content.byteOffset, file.content.byteLength);
+    }
+
+    if (file.gzip && typeof file.gzip.content !== 'string') {
+        file.gzip.content = Buffer.from(file.gzip.content.buffer, file.gzip.content.byteOffset, file.gzip.content.byteLength);
+    }
+
+    return file;
+}
+
+/**
+ * Spawns a background worker thread that lazily reads and processes every file
+ * in the directory, posting finished cache entries back to the main thread.
+ * This keeps the expensive first read (Sass/Babel/UglifyJS/gzip) off the event
+ * loop so requests don't stall. The worker is unref'd so it never keeps the
+ * process alive, and any failure simply leaves the synchronous read in place.
+ * @param {string} directory
+ * @param {object} options
+ * @param {object} processor
+ */
+function startWarmer(directory, options, processor) {
+    let worker;
+
+    try {
+        worker = new Worker(path.join(__dirname, 'warmer.js'), {
+            workerData: { directory, options }
+        });
+    } catch (err) {
+        //eslint-disable-next-line no-console
+        console.warn(`Electricity skipping cache warming:\n    ${err}`);
+        return;
+    }
+
+    worker.on('message', ({ file, urlPath }) => {
+        // Don't clobber an entry a request already populated via the
+        // synchronous fallback; both produce identical content.
+        if (!processor.files[urlPath]) {
+            processor.files[urlPath] = reviveFile(file);
+        }
+    });
+
+    worker.on('error', (err) => {
+        //eslint-disable-next-line no-console
+        console.warn(`Electricity cache warming error:\n    ${err}`);
+    });
+
+    // Never let warming hold the process open.
+    worker.unref();
+}
 
 exports.static = (directory, options) => {
     // Default to 'public' if the directory is not specified
@@ -73,329 +152,35 @@ exports.static = (directory, options) => {
         };
     }
 
-    // Create a local cache to hold the files
-    const files = {};
-
-    const snockets = new Snockets();
-
     let watcher;
 
     if (options.watch.enabled) {
         // Setup the watcher
         watcher = chokidar.watch(directory, { ignoreInitial: true });
+    }
 
+    const processor = createProcessor(directory, options, watcher);
+
+    if (watcher) {
         watcher.on('all', (eventName, filePath) => {
-            removeFile(filePath);
+            processor.removeFile(filePath);
         });
     }
 
-    /**
-     * Tries to read a file from local cache.
-     * Reads the file from disk if it's not present in the local cache.
-     * @param {string} urlPath
-     */
-    function fetchFile(urlPath) {
-        // Try to get the file from local cache
-        let file = files[urlPath];
-
-        // Return the file from cache if found
-        if (file) {
-            return file;
-        }
-
-        // Read the file from disk
-        file = readFile(urlPath);
-
-        // Put the file in local cache
-        files[urlPath] = file;
-
-        return file;
-    }
-
-    /**
-     * Converts a URL (/robots.txt) to a URL that includes the file's hash (/robots-3f54004ef6fc21b24a9e6069fc114fd9070b77a1.txt)
-     * @param {string} url
-     * @param {string} hash
-     */
-    function hashifyUrl(url, hash) {
-        if (!url.includes('.')) {
-            return url.replace(/([?#].*)?$/, `-${hash}$1`);
-        }
-
-        return url.replace(/\.([^.]*)([?#].*)?$/, `-${hash}.$1$2`);
-    }
-
-    /**
-     * Parses a URL path potentially containing a hash (/robots-3f54004ef6fc21b24a9e6069fc114fd9070b77a1.txt)
-     * into an object with a hash and path properties ({ hash: '3f54004ef6fc21b24a9e6069fc114fd9070b77a1', path: '/robots.txt' })
-     * @param {object} req
-     */
-    function parseUrlPath(urlPath) {
-        // https://regex101.com/r/j5hvRj/2
-        const regex = /\/.+(-([0-9a-f]{32,40}))/;
-        const matches = urlPath.match(regex);
-
-        if (!matches) {
-            return {
-                path: urlPath
-            };
-        }
-
-        return {
-            hash: matches[2],
-            path: urlPath.replace(matches[1], '')
-        };
-    }
-
-    function readCascadingStyleSheetsFile(filePath) {
-        let data;
-
-        // CSS
-        try {
-            data = fs.readFileSync(filePath).toString();
-        } catch (err) {
-            // Handle ENOENT (No such file or directory): https://nodejs.org/api/errors.html#common-system-errors
-            if (err.code !== 'ENOENT') {
-                throw err;
-            }
-
-            // SASS
-            const basename = path.basename(filePath, path.extname(filePath));
-            const sassFile = path.join(path.dirname(filePath), `${basename}.scss`);
-            const result = sass.compile(sassFile, options.sass);
-
-            data = result.css;
-
-            // SASS (watcher)
-            if (watcher) {
-                result.loadedUrls.forEach(file => {
-                    watcher.add(file.pathname);
-                });
-            }
-        }
-
-        // Update URLs in CSS: https://regex101.com/r/FxrppP/4
-        data = data.replace(/url\(['"]?(.*?)['"]?\)/g, (match, p1) => {
-            return `url(${urlBuilder(p1)})`;
-        });
-
-        // UglifyCSS
-        if (options.uglifycss.enabled) {
-            data = UglifyCss.processString(data, options.uglifycss);
-        }
-
-        return data;
-    }
-
-    function readFile(urlPath) {
-        let filePath = toFilePath(urlPath);
-        let extension = path.extname(filePath);
-        let data;
-
-        if (extension === '.css') {
-            data = readCascadingStyleSheetsFile(filePath);
-        } else if (extension === '.js') {
-            data = readJavaScriptFile(filePath);
-        } else {
-            data = fs.readFileSync(filePath);
-        }
-
-        const file = {
-            content: data,
-            contentLength: data.length,
-            contentType: mime.getType(urlPath),
-            hash: crypto.createHash('sha1').update(data).digest('hex')
-        };
-
-        // Don't gzip any content less that 1500 bytes (the size of a TCP packet). Only gzip specific content types.
-        if (options.gzip.enabled && file.contentLength > 1500 && gzipContentTypes.includes(file.contentType)) {
-            const gzipContent = zlib.gzipSync(file.content);
-
-            file.gzip = {
-                content: gzipContent,
-                contentLength: gzipContent.length
-            };
-        }
-
-        return file;
-    }
-
-    function readJavaScriptFile(filePath) {
-        let data;
-
-        // Snockets
-        try {
-            data = snockets.getConcatenation(filePath, options.snockets);
-        } catch(err) {
-            // Snockets can't parse, so just pass the js file along
-            //eslint-disable-next-line no-console
-            console.warn(`Snockets skipping ${filePath}:\n    ${err}`);
-        }
-
-        // Snockets (watcher)
-        if (watcher) {
-            try {
-                // Get all files in the snockets chain
-                const compiledChain = snockets.getCompiledChain(filePath, options.snockets);
-
-                // Add each file of the snockets chain to the watcher
-                compiledChain.forEach(c => {
-                    watcher.add(c.filename);
-                });
-            } catch(err) {
-                // Snockets can't parse, so skip watch
-                //eslint-disable-next-line no-console
-                console.warn(`Snockets skipping watch for ${filePath}:\n    ${err}`);
-            }
-        }
-
-        // If Snockets didn't parse the file, read it from disk
-        if (!data) {
-            data = fs.readFileSync(filePath).toString();
-        }
-
-        // Babel
-        try {
-            let result = babel.transformSync(data, {
-                ...options.babel,
-                presets: [require('@babel/preset-react')]
-            });
-
-            data = result.code;
-        } catch(err) {
-            // Babel can't transform, so just pass the file along
-            //eslint-disable-next-line no-console
-            console.warn(`Babel skipping ${filePath}:\n    ${err}`);
-        }
-
-        // UglifyJS
-        if (options.uglifyjs.enabled) {
-            const uglifyjsOptions = JSON.parse(JSON.stringify(options.uglifyjs));
-            delete uglifyjsOptions.enabled;
-
-            const result = UglifyJS.minify(data, uglifyjsOptions);
-
-            if (result.error) {
-                //eslint-disable-next-line no-console
-                console.warn(`UglifyJS skipping ${filePath}:\n    ${JSON.stringify(result.error)}`);
-            } else {
-                data = result.code;
-            }
-        }
-
-        return data;
-    }
-
-    /**
-     * Removes a file from the local cache.
-     * @param {*} filePath
-     */
-    function removeFile(filePath) {
-        let extension = path.extname(filePath);
-
-        if (extension === '.js') {
-            return removeJavaScriptFile(filePath);
-        } else if (extension === '.scss') {
-            return removeSassFile(filePath);
-        }
-
-        // Remove the changed file from the local cache
-        delete files[toUrlPath(filePath)];
-    }
-
-    /**
-     * Removes a JavaScript file from the local cache.
-     * @param {string} filePath
-     */
-     function removeJavaScriptFile(filePath) {
-        // Remove the changed file from the local cache
-        delete files[toUrlPath(filePath)];
-
-        // Resolve the absolute file path for the changed file
-        const absoluteFilePath = path.resolve(filePath);
-
-        // Find any parents that have a dependency on this file and remove them too
-        snockets.depGraph.parentsOf(absoluteFilePath).forEach(removeJavaScriptFile);
-    }
-
-    /**
-     * Removes a SASS file from the local cache.
-     * @param {string} filePath
-     */
-     function removeSassFile(filePath) {
-        const basename = path.basename(filePath, path.extname(filePath));
-        const cssFilePath = path.join(path.dirname(filePath), `${basename}.css`);
-        const urlPath = toUrlPath(cssFilePath);
-
-        // Remove the changed file from the local cache
-        delete files[urlPath];
-
-        // Resolve the absolute file path for the changed file
-        let absoluteFilePath = path.resolve(filePath);
-
-        // Try to resolve symlinks
-        try {
-            absoluteFilePath = fs.realpathSync(filePath);
-        } catch (e) {
-            // ignore error
-        }
-
-        const graph = sassGraph.parseDir(directory);
-        const sassFile = graph.index[absoluteFilePath];
-
-        if (sassFile) {
-            sassFile.importedBy.forEach(removeSassFile);
-        }
-    }
-
-    function urlBuilder(urlPath) {
-        let file;
-        const request = parseUrlPath(urlPath);
-        let url = urlPath;
-
-        try {
-            file = fetchFile(request.path);
-        } catch(err) {
-            // If we don't have a file that matches the specified URL path simply return the original URL path
-            return urlPath;
-        }
-
-        if (options.hashify) {
-            url = hashifyUrl(request.path, file.hash);
-        }
-
-        if (options.hostname) {
-            url = `https://${options.hostname}${url}`;
-        }
-
-        return url;
-    }
-
-    /**
-     * Converts a URL path (/robots.txt) to a file path (/Users/username/site/public/robots.txt).
-     * @param {string} urlPath
-     */
-    function toFilePath(urlPath) {
-        const myURL = new URL(urlPath, 'https://example.org/');
-        const pathname = myURL.pathname.replace(/^\//, '');
-        return path.resolve(directory, pathname);
-    }
-
-    /**
-     * Converts a file path (/Users/username/site/public/robots.txt) to a URL path (/robots.txt).
-     * @param {string} urlPath
-     */
-    function toUrlPath(filePath) {
-        const urlPath = path.posix.relative(directory, path.resolve(filePath));
-
-        return `/${urlPath}`;
+    // Warm the cache on a background thread so the first request for a file
+    // doesn't read and process it synchronously on the event loop. Watching
+    // keeps processing on the main thread (it registers discovered
+    // dependencies with the watcher), and options carrying functions can't be
+    // cloned into the worker, so warming is skipped in those cases.
+    if (!options.watch.enabled && !hasFunction(options)) {
+        startWarmer(directory, options, processor);
     }
 
     return function staticMiddleware(req, res, next) {
         // Register function in app.locals to help views build URLs: https://expressjs.com/en/api.html#app.locals
         if (req.app && !req.app.locals.electricity) {
             req.app.locals.electricity = {
-                url: urlBuilder
+                url: processor.urlBuilder
             };
         }
 
@@ -408,7 +193,7 @@ exports.static = (directory, options) => {
         const request = parseUrlPath(req.path);
 
         try {
-            file = fetchFile(request.path);
+            file = processor.fetchFile(request.path);
         } catch (err) {
             // Handle EISDIR (Is a directory): https://nodejs.org/api/errors.html#common-system-errors
             if (err.code === 'EISDIR') {
